[wip, #7267] added first test to detect denom-conflict on deposit
The tests have been extended to include conflict tests, see
test_exchange_api_conflicts.c

The first implemented test withdraws three coins with different
denominations, but using the same private key.  It deposits the first
coin successfully, but on deposit of the second coin it receives 409 and
the correct response details.

However, the test only seems to work for RSA, not CS!  I've disabled
test_exchange_conflicts_cs in the Makefile for now so that the builder
passes.

TBC...