activated tests for conflict wrt denom and age commitment
In src/testing/test_exchange_api_conflicts.c we test the conflicts when
two or more coins with the same private key have either different
denominations or different age restrictions.

Note that the test for different denominations does not fully work for
CS:  Even with the same master secret, the private keys also differ due
to the random choice of the nonce.

Fixes #7276