[withdraw] structured error handling
The ad-hoc error responses in the withdraw handler has been changed
to set a error type and potential witness data in the context, and
change the next phase to generate an error response.

Error types are also marked if they require a idempotency check (and
returning success, if so) before generating an error response.

Fixes #9377