[contrib] added doxygen and codespell checks to pre-commit
The git-hook for pre-commit has been extended to call
doxygen and codespell prior to accept a commit.

If no (significant) files have changed, exit early.

It calls the tools only when files were changed and the tools
codespell and doxygen are actually available.  It recommends to
install the tool, when it is missing.

Also, the files to skip for codespell (in CI) is now in a
seperate file contrib/ci/jobs/000-codespell/skip.txt.